# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Legal-KG API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/queue": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "List Queue",
        "operationId": "list_queue_api_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QueueItemResponse"
                  },
                  "title": "Response List Queue Api Queue Get"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "queue"
        ],
        "summary": "Clear Queue",
        "operationId": "clear_queue_api_queue_delete",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Clear Queue Api Queue Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/status": {
      "patch": {
        "tags": [
          "queue"
        ],
        "summary": "Update Case Status",
        "operationId": "update_case_status_api_cases__case_id__status_patch",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/review": {
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Record Review Decision",
        "operationId": "record_review_decision_api_cases__case_id__review_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "queue"
        ],
        "summary": "Update Legacy Review",
        "operationId": "update_legacy_review_api_cases__case_id__review_patch",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/audit": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "Get Case Audit",
        "operationId": "get_case_audit_api_cases__case_id__audit_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntryResponse"
                  },
                  "title": "Response Get Case Audit Api Cases  Case Id  Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuditEntryResponse": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "actor": {
            "type": "string",
            "title": "Actor",
            "default": "operator"
          },
          "old_value": {
            "type": "string",
            "title": "Old Value",
            "default": ""
          },
          "new_value": {
            "type": "string",
            "title": "New Value",
            "default": ""
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "action"
        ],
        "title": "AuditEntryResponse"
      },
...
      "DecisionReviewRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "accepted",
              "corrected",
              "rejected"
            ],
            "title": "Decision"
          },
          "note": {
            "type": "string",
            "maxLength": 1000,
            "title": "Note",
            "default": ""
          },
          "corrected_label": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "dung",
                  "hieu_lam",
                  "can_kiem_chung"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Label"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DecisionReviewRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NhanNguon": {
        "type": "string",
        "enum": [
          "co_nguon_xac_nhan",
          "co_bac_bo_chinh_thuc",
          "chua_tim_thay_nguon"
        ],
        "title": "NhanNguon"
      },
      "NhanPhanLoai": {
        "type": "string",
        "enum": [
          "dung",
          "hieu_lam",
          "can_kiem_chung"
        ],
        "title": "NhanPhanLoai"
      },
...
      "QueueItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "comment_id": {
            "type": "string",
            "title": "Comment Id",
            "default": ""
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": ""
          },
          "claim": {
            "type": "string",
            "title": "Claim"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citations"
          },
          "label": {
            "$ref": "#/components/schemas/NhanPhanLoai"
          },
          "source_label": {
            "$ref": "#/components/schemas/NhanNguon"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 0
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "default": "Forum"
          },
          "account": {
            "type": "string",
            "title": "Account",
            "default": "Ngu\u1ed3n ch\u01b0a x\u00e1c \u0111\u1ecbnh"
          },
          "published_at": {
            "type": "string",
            "title": "Published At",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          },
          "reach": {
            "type": "integer",
            "title": "Reach",
            "default": 0
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "new"
          },
          "document": {
            "type": "string",
            "title": "Document",
            "default": "Ngh\u1ecb \u0111\u1ecbnh 174/2026/N\u0110-CP"
          },
          "provision": {
            "type": "string",
            "title": "Provision",
            "default": ""
          },
          "penalty": {
            "type": "string",
            "title": "Penalty",
            "default": ""
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "default": "Ch\u01b0a x\u00e1c \u0111\u1ecbnh"
          },
          "source_title": {
            "type": "string",
            "title": "Source Title",
            "default": ""
          },
          "source_url": {
            "type": "string",
            "title": "Source Url",
            "default": ""
          },
          "source_agency": {
            "type": "string",
            "title": "Source Agency",
            "default": ""
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "default": 50
          },
          "confidence": {
            "type": "integer",
            "title": "Confidence",
            "default": 50
          },
          "spread_risk": {
            "type": "integer",
            "title": "Spread Risk",
            "default": 0
          },
          "ai_accuracy": {
            "type": "integer",
            "title": "Ai Accuracy",
            "default": 0
          },
          "source_reliability": {
            "type": "integer",
            "title": "Source Reliability",
            "default": 0
          },
          "human_label": {
            "type": "string",
            "title": "Human Label",
            "default": ""
          },
          "human_source_label": {
            "type": "string",
            "title": "Human Source Label",
            "default": ""
          },
          "reviewer_notes": {
            "type": "string",
            "title": "Reviewer Notes",
            "default": ""
          },
          "comments": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Comments"
          },
          "reviewer_label": {
            "type": "string",
            "title": "Reviewer Label",
            "default": ""
          },
          "reviewer_reason": {
            "type": "string",
            "title": "Reviewer Reason",
            "default": ""
          },
          "reviewer_note": {
            "type": "string",
            "title": "Reviewer Note",
            "default": ""
          },
          "reviewed_at": {
            "type": "string",
            "title": "Reviewed At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "claim",
          "label",
          "source_label",
          "reason"
        ],
        "title": "QueueItemResponse"
      },
      "ReviewRequest": {
        "properties": {
          "human_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Label"
          },
          "human_source_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Source Label"
          },
          "reviewer_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Notes"
          },
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          }
        },
        "type": "object",
        "title": "ReviewRequest"
      },
      "StatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewer_label": {
            "type": "string",
            "title": "Reviewer Label",
            "default": ""
          },
          "reviewer_reason": {
            "type": "string",
            "title": "Reviewer Reason",
            "default": ""
          },
          "reviewer_note": {
            "type": "string",
            "title": "Reviewer Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}